third_party/WebKit/LayoutTests/flag-specific/enable-blink-features=LayoutNG/external/wpt/css/css-shapes/spec-examples/shape-outside-010-expected.txt

This is a testharness.js-based test.
FAIL CSS Test: Shape from image - url(), alpha channel, opacity 0 Uncaught Error: assert_approx_equals: Line 0 is positioned properly expected 190 +/- 2 but got 180
Harness: the test ran to completion.
